+++ terrace/terracetree.cpp
@@ -406,7 +406,7 @@ void TerraceTree::fillbrNodes(){
     NodeVector br_1, br_2;
     getBranches(br_1, br_2);
     int id;
-    for(int i=0; i<branchNum; i++){
+    for(size_t i=0; i<branchNum; i++){
         id=br_1[i]->findNeighbor(br_2[i])->id;
         brNodes[id]={br_1[i],br_2[i]};;
         /*for(const auto& b: brNodes[id]){
